Programmatic Completion Enforcement (#16)
Add a way to enforce that the llm can't say they are complete until they satisfy a set of validations.

Each validator (currently just the certora prover) must detect if it is happy with the entire spec (in the case of the prover, this means all rules pass) and then record a hash of the current VFS in the state.

Within the result tool, the required validations (recorded in the context) are checked against the hash of the current VFS. If there is a mismatch (or the validation hasn't even been run) then the result tool is rejected, and the LLM is instructed to get back on track (by creating a TODO list).